% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{oja_disconnect}
\alias{oja_disconnect}
\title{Chiude una connessione DuckDB OJA}
\usage{
oja_disconnect(con)
}
\arguments{
\item{con}{Connessione restituita da \code{\link[=oja_connect]{oja_connect()}}.}
}
\value{
Invisibilmente \code{TRUE}.
}
\description{
Equivalente a \code{DBI::dbDisconnect(con, shutdown = TRUE)}. Restituisce
invisibilmente \code{TRUE}.
}
