% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{oja_sftp_download}
\alias{oja_sftp_download}
\title{Scarica un singolo file via SFTP}
\usage{
oja_sftp_download(remote_name, local_path, config = oja_sftp_config())
}
\arguments{
\item{remote_name}{Nome del file remoto (relativo a \code{config$remote_dir}).}

\item{local_path}{Percorso locale di destinazione.}

\item{config}{Output di \code{\link[=oja_sftp_config]{oja_sftp_config()}}.}
}
\value{
Invisibilmente \code{local_path}.
}
\description{
Download atomico: scrive in \verb{<local_path>.part} e rinomina al termine.
Errore lasciato propagare al chiamante; in caso di errore il file \code{.part}
resta sul disco come traccia diagnostica.
}
