% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{oja_skill_dim}
\alias{oja_skill_dim}
\title{Vista lazy su \code{dim_esco_skill}}
\usage{
oja_skill_dim(con)
}
\arguments{
\item{con}{Connessione restituita da \code{\link[=oja_connect]{oja_connect()}}.}
}
\value{
Un \code{tbl_lazy} di \code{dbplyr}.
}
\description{
Vista lazy su \code{dim_esco_skill}
}
